Corso Integrativo Cloud Computing/Codice/04/photogallery_v6/requirements.txt
boto3==1.10.14
botocore==1.13.14
Click==7.0
docutils==0.15.2
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
Pillow==6.2.1
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
